Made fix to keep download_process_data()/pipeline_astrometry() from stopping when an error is thrown. Removed symlinking to e11 fits files in working directory if e91 fits files found in working directory.

+++ neoexchange/photometrics/catalog_subs.py
@@ -1299,8 +1299,15 @@ def sort_rocks(fits_files):
             if not os.path.exists(object_directory):
                 os.makedirs(object_directory)
             dest_filepath = os.path.join(object_directory, os.path.basename(fits_filepath))
-            if not os.path.exists(dest_filepath):
-                os.symlink(fits_filepath, dest_filepath)
+            #if the file is an e91 and an e11 exists in the working directory, remove the link to the e11 and link the e91
+            if 'e91' in fits_filepath and os.path.exists(dest_filepath.replace('e91.fits', 'e11.fits')):
+                os.unlink(dest_filepath.replace('e91.fits', 'e11.fits'))
+                if not os.path.exists(dest_filepath):
+                    os.symlink(fits_filepath, dest_filepath)
+            #if the file is an e11 and an e91 doesn't exit in the working directory, create link to the e11
+            elif 'e11' in fits_filepath and not os.path.exists(dest_filepath.replace('e11.fits', 'e91.fits')):
+                if not os.path.exists(dest_filepath):
+                    os.symlink(fits_filepath, dest_filepath)
     return objects
 
 def find_first_last_frames(fits_files):
